+++ index.ts
@@ -3,6 +3,7 @@ export type Graph = [Set<Vertex>, Set<Arc>];
 export interface ParseResult {
 	graph: Graph;
 	root: Vertex;
+	rootWeight: number;
 }
 export interface Vertex {
 	label?: string;
@@ -11,35 +12,54 @@ export function parse(s: string): ParseResult {
 	if (typeof s !== "string") {
 		throw new Error(`Not a string: ${s}`);
 	}
-	s = s.trim();
 	const minLength = /;$/.test(s) ? 2 : 1;
 	if (s.length < minLength) {
 		throw new Error(`Not long enough to be a Newick string: "${s}".`);
 	}
 	const graph: Graph = [new Set<Vertex>(), new Set<Arc>()];
-	const root = readVertex(new CharBuffer(s), graph);
+	const buffer = new CharBuffer(s);
+	const root = readVertex(buffer, graph);
 	return {
 		graph,
 		root,
+		rootWeight: readRootWeight(buffer),
 	};
 }
 export function write(graph: Graph): string {
+	if (!Array.isArray(graph)) {
+		throw new Error(`Not an array: ${String(graph) || "<empty>"}`);
+	}
+	if (graph.length < 2 || graph.slice(0, 2).some(element => !(element instanceof Set))) {
+		throw new Error(`Not a graph: [${graph.join(", ")}]`);
+	}
+	if (Array.from(graph[0]).some(vertex => typeof vertex !== "object")) {
+		throw new Error("Invalid vertex set.");
+	}
+	if (Array.from(graph[1]).some(arc => !Array.isArray(arc) || arc.length < 3 || typeof arc[0] !== "object" || typeof arc[1] !== "object" || typeof arc[2] !== "number")) {
+		throw new Error("Invalid arc set.");
+	}
 	return new GraphWriter(graph).write();
 }
 class CharBuffer {
 	constructor(private readonly string = "") {
 		this.length = this.string.length;
 	}
-	atEnd() {
+	public get rest() {
+		if (this.atEnd()) {
+			throw new Error("End of buffer reached.");
+		}
+		return this.string.substr(this.pos);
+	}
+	public atEnd() {
 		return this.pos >= this.length;
 	}
-	back() {
+	public back() {
 		if (this.pos <= 0) {
 			throw new Error("At start of buffer.");
 		}
 		this.pos--;
 	}
-	read() {
+	public read() {
 		if (this.atEnd()) {
 			throw new Error("End of buffer reached.");
 		}
@@ -126,6 +146,20 @@ function findVertexWithLabel(label: string, vertices: Set<Vertex>) {
 	}
 	return null;
 }
+function readRootWeight(buffer: CharBuffer): number {
+	if (!buffer.atEnd()) {
+		if (buffer.read() === ":") {
+			const weight = readWeight(buffer);
+			if (buffer.atEnd()) {
+				return weight;
+			}
+		} else {
+			buffer.back();
+		}
+		throw new Error(`Extra content beyond end of Newick tree: "${buffer.rest}".`);
+	}
+	return NaN;
+}
 function readVertex(buffer: CharBuffer, graph: Graph): Vertex {
 	let token: string;
 	do {
@@ -183,7 +217,7 @@ function readVertexLabel(buffer: CharBuffer, vertices: Set<Vertex>): Vertex {
 				buffer.back();
 				break;
 			}
-			if (label.length !== 0 || !(/^\s$/.test(token))) {
+			if (label.length !== 0 || !(/\s/.test(token))) {
 				label += token;
 			}
 		}
@@ -192,10 +226,10 @@ function readVertexLabel(buffer: CharBuffer, vertices: Set<Vertex>): Vertex {
 		return {};
 	}
 	label = label
-		.replace(/;$/, '')
-		.replace(/^\s+/, '')
-		.replace(/\s+$/, '')
-		.replace(/\s\s+/g, ' ');
+		.trim()
+		.replace(/;$/, "")
+		.replace(/\s+$/, "")
+		.replace(/\s/g, " ");
 	if (label.length === 0) {
 		return {};
 	}